proposal to organize tokenize logic in its own mod
