+++ zcashrpc-typegen/src/main.rs
@@ -3,6 +3,7 @@
 
 mod error;
 mod special_cases;
+mod tokenize;
 use error::TypegenResult;
 use proc_macro2::TokenStream;
 use quote::quote;
@@ -205,7 +206,7 @@ fn enumgen(
                 }
                 non_object => {
                     let (variant_body_tokens, new_acc, _terminal_enum) =
-                        tokenize_value(
+                        tokenize::value(
                             &variant_name,
                             non_object,
                             acc.clone(),
@@ -302,7 +303,7 @@ fn handle_fields(
     for (mut field_name, val) in inner_nodes {
         //special case handling
         if &field_name == "xxxx" {
-            new_code = tokenize_value(struct_name, val, Vec::new(), false)?.1; // .0 unused
+            new_code = tokenize::value(struct_name, val, Vec::new(), false)?.1; // .0 unused
             case = special_cases::Case::FourXs;
             break;
         }
@@ -318,7 +319,7 @@ fn handle_fields(
 
         //temp_acc needed because destructuring assignments are unstable
         //see https://github.com/rust-lang/rust/issues/71126 for more info
-        let (mut tokenized_val, temp_acc, _terminal_enum) = tokenize_value(
+        let (mut tokenized_val, temp_acc, _terminal_enum) = tokenize::value(
             &capitalize_first_char(&field_name),
             val,
             new_code,
@@ -353,53 +354,13 @@ fn handle_fields(
     })
 }
 
-fn handle_terminal_enum(
-    label: &str,
-    name: &str,
-    called_by_alias: bool,
-) -> TokenStream {
-    let variants = label
-        .strip_prefix("ENUM:")
-        .unwrap()
-        .split(',')
-        .map(|x| x.trim());
-    let variant_idents = variants
-        .clone()
-        .map(|x| {
-            proc_macro2::TokenTree::Ident(callsite_ident(
-                &x.split('-').map(capitalize_first_char).collect::<String>(),
-            ))
-            .into()
-        })
-        .collect::<Vec<TokenStream>>();
-    let variant_idents_renames = variants
-        .map(|x| format!("#[serde(rename = \"{}\")]", x).parse().unwrap())
-        .collect::<Vec<TokenStream>>();
-    #[rustfmt::skip]
-    let name_tokens = callsite_ident( 
-        &(
-            if called_by_alias {
-                format!("{}Response", name)
-            } else {
-                name.to_string()
-            }
-        )
-    );
-    quote!(
-        #[derive(Debug, serde::Deserialize, serde::Serialize)]
-        pub enum #name_tokens {
-            #(#variant_idents_renames #variant_idents,)*
-        }
-    )
-}
-
 fn alias(
     data: serde_json::Value,
     name: &str,
     acc: Vec<TokenStream>,
 ) -> TypegenResult<Vec<TokenStream>> {
     let ident = callsite_ident(&name);
-    let (type_body, mut acc, terminal_enum) = tokenize_value(
+    let (type_body, mut acc, terminal_enum) = tokenize::value(
         &capitalize_first_char(name.trim_end_matches("Response")),
         data,
         acc,
@@ -412,103 +373,6 @@ fn alias(
         acc.push(aliased);
     }
     Ok(acc)
-}
-
-fn tokenize_value(
-    name: &str,
-    val: serde_json::Value,
-    acc: Vec<TokenStream>,
-    called_by_alias: bool,
-) -> TypegenResult<(TokenStream, Vec<TokenStream>, bool)> {
-    match val {
-        serde_json::Value::String(label) => {
-            tokenize_terminal(name, label.as_str(), acc, called_by_alias)
-        }
-        serde_json::Value::Array(vec) => {
-            tokenize_array(name, vec, acc).map(|x| (x.0, x.1, false))
-        }
-        serde_json::Value::Object(obj) => {
-            tokenize_object(name, obj, acc).map(|x| (x.0, x.1, false))
-        }
-        otherwise => Err(error::QuizfaceAnnotationError {
-            kind: error::InvalidAnnotationKind::from(otherwise),
-            location: name.to_string(),
-        })?,
-    }
-}
-
-fn tokenize_terminal(
-    name: &str,
-    label: &str,
-    mut acc: Vec<TokenStream>,
-    called_by_alias: bool,
-) -> TypegenResult<(TokenStream, Vec<TokenStream>, bool)> {
-    Ok((
-        match label {
-            "Decimal" => quote!(rust_decimal::Decimal),
-            "bool" => quote!(bool),
-            "String" => quote!(String),
-            "hexadecimal" => quote!(String),
-            "INSUFFICIENT" => quote!(compile_error!(
-                "Insufficient zcash-cli help output to autogenerate type"
-            )),
-            enumeration if enumeration.starts_with("ENUM:") => {
-                let ident = callsite_ident(name);
-                acc.push(handle_terminal_enum(
-                    enumeration,
-                    name,
-                    called_by_alias,
-                ));
-                return Ok((quote!(#ident), acc, true));
-            }
-            otherwise => {
-                return Err(error::QuizfaceAnnotationError {
-                    kind: error::InvalidAnnotationKind::from(
-                        serde_json::Value::String(otherwise.to_string()),
-                    ),
-                    location: name.to_string(),
-                }
-                .into())
-            }
-        },
-        acc,
-        false,
-    ))
-}
-
-fn tokenize_array(
-    name: &str,
-    mut array_of: Vec<serde_json::Value>,
-    acc: Vec<TokenStream>,
-) -> TypegenResult<(TokenStream, Vec<TokenStream>)> {
-    let (val, acc, _terminal_enum) = tokenize_value(
-        name,
-        array_of.pop().ok_or(error::QuizfaceAnnotationError {
-            kind: error::InvalidAnnotationKind::EmptyArray,
-            location: name.to_string(),
-        })?,
-        acc,
-        false,
-    )?;
-    Ok((quote!(Vec<#val>), acc))
-}
-
-fn tokenize_object(
-    name: &str,
-    val: serde_json::Map<String, serde_json::Value>,
-    acc: Vec<TokenStream>,
-) -> TypegenResult<(TokenStream, Vec<TokenStream>)> {
-    let ident = callsite_ident(name);
-    let (case, acc) = structgen(val, name, acc)?;
-    match case {
-        special_cases::Case::Regular => Ok((quote!(#ident), acc)),
-        special_cases::Case::FourXs => {
-            Ok((quote!(std::collections::HashMap<String, #ident>), acc))
-        }
-        otherwise => {
-            panic!("structgen should not return variant {:?}", otherwise)
-        }
-    }
 }
 
 fn capitalize_first_char(input: &str) -> String {
@@ -524,7 +388,7 @@ mod unit {
         use crate::*;
         #[test]
         fn tokenize_value_string() {
-            let quoted_string = tokenize_value(
+            let quoted_string = tokenize::value(
                 "some_field",
                 serde_json::json!("String"),
                 Vec::new(),
@@ -537,7 +401,7 @@ mod unit {
         }
         #[test]
         fn tokenize_value_number() {
-            let quoted_number = tokenize_value(
+            let quoted_number = tokenize::value(
                 "some_field",
                 serde_json::json!("Decimal"),
                 Vec::new(),
@@ -550,7 +414,7 @@ mod unit {
         }
         #[test]
         fn tokenize_value_bool() {
-            let quoted_bool = tokenize_value(
+            let quoted_bool = tokenize::value(
                 "some_field",
                 serde_json::json!("bool"),
                 Vec::new(),
@@ -577,7 +441,7 @@ mod unit {
         }
         #[test]
         fn tokenize_object_simple_unnested() {
-            let quoted_object = tokenize_value(
+            let quoted_object = tokenize::value(
                 "somefield",
                 serde_json::json!(
                     {
